lab1-mcp/pyproject.toml
[project]
name = "langchain-k8s-mcp"
version = "0.1.0"
description = "Gemini LLM integrated with Kubernetes MCP server"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "python-dotenv",
    "requests",
]